+++ src/agent/main/java/org/comroid/mcsd/agent/AgentRunner.java
@@ -9,13 +9,15 @@ import org.comroid.api.DelegateStream;
 import org.comroid.api.Event;
 import org.comroid.api.Polyfill;
 import org.comroid.mcsd.agent.controller.ConsoleController;
-import org.comroid.mcsd.api.model.Status;
 import org.comroid.mcsd.core.entity.*;
-import org.comroid.mcsd.core.module.*;
+import org.comroid.mcsd.core.module.console.ConsoleModule;
+import org.comroid.mcsd.core.module.shell.BackupModule;
+import org.comroid.mcsd.core.module.shell.ExecutionModule;
+import org.comroid.mcsd.core.module.shell.UpdateModule;
+import org.comroid.mcsd.core.module.status.StatusModule;
 import org.comroid.mcsd.core.repo.ServerRepo;
 import org.comroid.mcsd.core.repo.ShRepo;
 import org.comroid.mcsd.util.Utils;
-import org.comroid.util.StandardValueType;
 import org.comroid.util.Streams;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
