+++ src/agent/main/java/org/comroid/mcsd/agent/controller/ConsoleController.java
@@ -11,8 +11,7 @@ import org.comroid.mcsd.agent.config.WebSocketConfig;
 import org.comroid.mcsd.core.entity.Server;
 import org.comroid.mcsd.core.entity.UserData;
 import org.comroid.mcsd.core.model.ServerConnection;
-import org.comroid.mcsd.core.module.ConsoleModule;
-import org.comroid.mcsd.core.module.ExecutionModule;
+import org.comroid.mcsd.core.module.shell.ExecutionModule;
 import org.comroid.mcsd.core.repo.UserDataRepo;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
