+++ src/core/main/java/org/comroid/mcsd/core/module/shell/BackupModule.java
@@ -1,4 +1,4 @@
-package org.comroid.mcsd.core.module;
+package org.comroid.mcsd.core.module.shell;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -11,6 +11,9 @@ import org.comroid.api.io.FileHandle;
 import org.comroid.mcsd.api.model.Status;
 import org.comroid.mcsd.core.entity.Backup;
 import org.comroid.mcsd.core.entity.Server;
+import org.comroid.mcsd.core.module.console.ConsoleModule;
+import org.comroid.mcsd.core.module.ServerModule;
+import org.comroid.mcsd.core.module.status.StatusModule;
 import org.comroid.mcsd.core.repo.BackupRepo;
 import org.comroid.util.Archiver;
 import org.comroid.util.PathUtil;
@@ -18,8 +21,6 @@ import org.comroid.util.Stopwatch;
 
 import java.io.File;
 import java.nio.file.Paths;
-import java.time.Duration;
-import java.time.Instant;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
