+++ src/core/main/java/org/comroid/mcsd/core/module/status/UptimeModule.java
@@ -1,22 +1,18 @@
-package org.comroid.mcsd.core.module;
+package org.comroid.mcsd.core.module.status;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.java.Log;
-import org.apache.commons.lang3.NotImplementedException;
 import org.comroid.api.Component;
-import org.comroid.api.Event;
 import org.comroid.api.os.OS;
-import org.comroid.mcsd.api.model.IStatusMessage;
-import org.comroid.mcsd.api.model.Status;
 import org.comroid.mcsd.core.entity.Server;
 import org.comroid.mcsd.core.entity.ServerUptimeEntry;
-import org.comroid.mcsd.core.repo.ServerRepo;
+import org.comroid.mcsd.core.module.shell.ExecutionModule;
+import org.comroid.mcsd.core.module.ServerModule;
 import org.comroid.mcsd.core.repo.ServerUptimeRepo;
 
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 
 import static java.time.Instant.now;
